WIP: remove ValueActionFactory
